refactor(demo): share the account card behind a chain adapter
Replace the presentational shell (28 flat props re-plumbed by every chain
card) with a shared ChainAccountCard that owns the card state machine —
polling, funding gate, send lifecycle, Max — and a small ChainAdapter that
carries the genuinely chain-specific labels and behavior. The chain cards
shrink to adapter definitions focused on validation, fees, and signing.
